Remove reveal-multiplex
GitHub warned of a security vulnerability in the socket.io version used
by the muliplex plugin.

It turns out this plugin solves a problem we don't have. From the
reveal-multiplex readme[1]:

"The multiplex plugin allows your audience to view the slides of the
reveal.js presentation you are controlling on their own phone, tablet or
laptop."

[1] https://github.com/reveal/multiplex

It looks like this plugin was in the original repository we used a
template for our slides.

I suggest we remove the plugin rather than try to update it or its
dependencies.